improved KuveyPos and InterPos
